Split worker process tests by owning module
