+++ src/worker_process.rs
@@ -1,45 +1,20 @@
 use std::path::PathBuf;
-#[cfg(test)]
-use std::sync::atomic::Ordering;
 use std::sync::{Arc, Mutex, atomic::AtomicBool};
-#[cfg(test)]
-use std::thread;
 use std::time::Duration;
 
 use crate::backend::{Backend, WorkerLaunch};
 use crate::completion_reply::CompletionInfo;
-#[cfg(test)]
-use crate::ipc::IpcEchoEvent;
-#[cfg(test)]
-use crate::output_capture::OutputTextSource;
 use crate::output_capture::{
     OUTPUT_RING_CAPACITY_BYTES, OutputBuffer, OutputTimeline, ensure_output_ring,
     reset_last_reply_marker_offset, reset_output_ring,
 };
-#[cfg(test)]
-use crate::output_snapshot::{SnapshotWithImages, snapshot_page_with_images};
-#[cfg(test)]
-use crate::output_timeline::{EchoCollapseMode, collapse_echo_with_attribution};
 use crate::oversized_output::OversizedOutputMode;
-#[cfg(test)]
-use crate::pager;
 use crate::pager::Pager;
 use crate::pending_output_tape::PendingOutputTape;
-#[cfg(test)]
-use crate::pending_output_tape::PendingSidebandKind;
-#[cfg(test)]
-use crate::reply_presentation::trim_echo_prefix_after_leading_nonstdout_contents;
-#[cfg(test)]
-use crate::reply_presentation::{
-    maybe_trim_echo_prefix, should_trim_echo_prefix, trim_echo_then_append_protocol_warnings,
-    trim_matching_echo_event_suffix_from_contents,
-};
 use crate::sandbox::{SandboxState, SandboxStateUpdate};
 use crate::sandbox_cli::SandboxCliPlan;
 pub(crate) use crate::stdin_payload::{WriteStdinControlAction, split_write_stdin_control_prefix};
 use crate::worker_protocol::WorkerReply;
-#[cfg(test)]
-use crate::worker_protocol::{ContentOrigin, WorkerContent, WorkerErrorCode};
 use crate::worker_supervisor::{GuardrailEvent, GuardrailShared, WorkerProcess};
 
 mod backend_driver;
@@ -54,6 +29,8 @@ mod restart;
 mod sandbox_state;
 mod session_lifecycle;
 mod session_reset_reply;
+#[cfg(test)]
+mod test_support;
 mod worker_launch;
 mod write_dispatch;
 mod write_flow;
@@ -61,8 +38,6 @@ mod write_preflight;
 
 use self::backend_driver::{BackendDriver, new_backend_driver};
 use self::output_state::PrefixCapture;
-#[cfg(test)]
-use self::reply_state::mark_busy_follow_up_reply;
 use self::request_lifecycle::RequestState;
 pub(crate) use self::write_flow::WriteStdinOptions;
 
@@ -359,425 +334,8 @@ impl WorkerManager {
 
 #[cfg(test)]
 mod tests {
+    use super::test_support::{cwd_test_mutex, env_test_mutex};
     use super::*;
-    use crate::ipc::WorkerToServerIpcMessage;
-    #[cfg(any(target_family = "unix", target_family = "windows"))]
-    use crate::ipc::{IpcHandlers, IpcPlotImage};
-    use crate::output_capture::{
-        OUTPUT_RING_CAPACITY_BYTES, OutputEventKind, OutputRange, OutputRing, OutputTextSpan,
-        ensure_output_ring, reset_last_reply_marker_offset, reset_output_ring,
-    };
-    use crate::pending_output_tape::PendingOutputEvent;
-    use crate::pending_output_tape::PendingTextSource;
-    use crate::sandbox::SandboxPolicy;
-    #[cfg(target_os = "linux")]
-    use crate::sandbox::sandbox_state_update_from_codex_meta;
-    use crate::sandbox_cli::resolve_effective_sandbox_state_with_defaults;
-    use crate::worker_protocol::TextStream;
-    #[cfg(target_family = "unix")]
-    use crate::worker_supervisor::capture_recorded_unix_kills;
-    #[cfg(target_os = "linux")]
-    use crate::worker_supervisor::linux_sandbox_startup_retryable;
-    use crate::worker_supervisor::{
-        LiveOutputCapture, apply_debug_startup_env, cleanup_worker_session_tmpdir,
-        persist_worker_startup_log,
-    };
-    #[cfg(target_family = "windows")]
-    use crate::worker_supervisor::{
-        WINDOWS_IPC_CONNECT_MAX_WAIT, handle_windows_ipc_connect_result, request_soft_termination,
-    };
-    use base64::Engine as _;
-    #[cfg(target_os = "linux")]
-    use serde_json::json;
-    use std::process::{Child, Command};
-    use std::sync::mpsc;
-    use std::sync::{Mutex, MutexGuard, OnceLock};
-
-    fn cwd_test_mutex() -> &'static Mutex<()> {
-        static TEST_MUTEX: OnceLock<Mutex<()>> = OnceLock::new();
-        TEST_MUTEX.get_or_init(|| Mutex::new(()))
-    }
-
-    fn env_test_mutex() -> &'static Mutex<()> {
-        static TEST_MUTEX: OnceLock<Mutex<()>> = OnceLock::new();
-        TEST_MUTEX.get_or_init(|| Mutex::new(()))
-    }
-
-    fn output_ring_test_guard() -> MutexGuard<'static, ()> {
-        crate::output_capture::output_ring_test_mutex()
-            .lock()
-            .unwrap_or_else(|err| err.into_inner())
-    }
-
-    #[cfg(target_family = "unix")]
-    fn worker_process_test_temp_parent(label: &str) -> PathBuf {
-        let root = std::env::temp_dir()
-            .join("mcp-repl-test-scratch")
-            .join(label);
-        std::fs::create_dir_all(&root).expect("create worker process test temp parent");
-        root
-    }
-
-    #[test]
-    fn python_backend_prepares_windows_sandbox_launch() {
-        assert!(
-            worker_launch::backend_prepares_windows_sandbox_launch(Backend::Python),
-            "Python uses the embedded worker wrapper and needs the prepared Windows capability SID"
-        );
-    }
-
-    fn echo_event(prompt: &str, line: &str) -> IpcEchoEvent {
-        IpcEchoEvent {
-            prompt: prompt.to_string(),
-            line: line.to_string(),
-            source: OutputTextSource::Ipc,
-        }
-    }
-
-    fn contents_text(contents: &[WorkerContent]) -> String {
-        contents
-            .iter()
-            .filter_map(|content| match content {
-                WorkerContent::ContentText { text, .. } => Some(text.as_str()),
-                _ => None,
-            })
-            .collect::<Vec<_>>()
-            .join("")
-    }
-
-    fn pager_buffer_from_worker_text(text: &str) -> crate::pager::PagerBuffer {
-        pager_buffer_from_worker_text_with_source_end(text, text.len() as u64)
-    }
-
-    fn static_pager_buffer_from_worker_text(text: &str) -> crate::pager::PagerBuffer {
-        pager_buffer_from_worker_text_with_source_end(text, u64::MAX)
-    }
-
-    fn pager_buffer_from_worker_text_with_source_end(
-        text: &str,
-        source_end: u64,
-    ) -> crate::pager::PagerBuffer {
-        crate::pager::PagerBuffer::from_bytes_and_events(
-            text.as_bytes().to_vec(),
-            Vec::new(),
-            vec![OutputTextSpan {
-                start_byte: 0,
-                end_byte: text.len(),
-                is_stderr: false,
-                origin: ContentOrigin::Worker,
-                source: crate::output_capture::OutputTextSource::Raw,
-            }],
-            source_end,
-        )
-    }
-
-    #[cfg(target_family = "unix")]
-    fn sleeping_test_child() -> Child {
-        Command::new("sh")
-            .args(["-c", "sleep 30"])
-            .spawn()
-            .expect("spawn sleeping test child")
-    }
-
-    #[cfg(target_family = "windows")]
-    fn sleeping_test_child() -> Child {
-        Command::new("powershell.exe")
-            .args(["-NoProfile", "-Command", "Start-Sleep -Seconds 30"])
-            .spawn()
-            .expect("spawn sleeping test child")
-    }
-
-    #[cfg(target_family = "unix")]
-    fn successful_test_child() -> Child {
-        Command::new("sh")
-            .args(["-c", "exit 0"])
-            .spawn()
-            .expect("spawn exiting test child")
-    }
-
-    #[cfg(target_family = "windows")]
-    fn successful_test_child() -> Child {
-        Command::new("powershell.exe")
-            .args(["-NoProfile", "-Command", "exit 0"])
-            .spawn()
-            .expect("spawn exiting test child")
-    }
-
-    #[cfg(target_family = "unix")]
-    fn failing_test_status() -> std::process::ExitStatus {
-        Command::new("sh")
-            .args(["-c", "exit 7"])
-            .status()
-            .expect("collect failing exit status")
-    }
-
-    #[cfg(target_family = "unix")]
-    fn test_worker_process(child: Child) -> WorkerProcess {
-        WorkerProcess::new_for_test(child)
-    }
-
-    #[cfg(target_family = "windows")]
-    fn test_worker_process(child: Child) -> WorkerProcess {
-        WorkerProcess::new_for_test(child)
-    }
-
-    #[test]
-    fn trims_echo_prefix_across_text_chunks() {
-        let mut contents = vec![
-            WorkerContent::stdout("> x <- 1\n"),
-            WorkerContent::stdout("> y <- 2\n[1] 2\n"),
-        ];
-        maybe_trim_echo_prefix(&mut contents, Some("> x <- 1\n> y <- 2\n"), true);
-        let text = match &contents[0] {
-            WorkerContent::ContentText { text, .. } => text.as_str(),
-            _ => "",
-        };
-        assert_eq!(text, "[1] 2\n");
-    }
-
-    #[test]
-    fn does_not_trim_on_mismatch() {
-        let mut contents = vec![WorkerContent::stdout("> x <- 1\n[1] 1\n")];
-        maybe_trim_echo_prefix(&mut contents, Some("> y <- 2\n"), true);
-        let text = match &contents[0] {
-            WorkerContent::ContentText { text, .. } => text.as_str(),
-            _ => "",
-        };
-        assert_eq!(text, "> x <- 1\n[1] 1\n");
-    }
-
-    #[test]
-    fn does_not_trim_when_leading_stderr() {
-        let mut contents = vec![
-            WorkerContent::stderr("stderr: boom\n"),
-            WorkerContent::stdout("> x <- 1\n[1] 1\n"),
-        ];
-        maybe_trim_echo_prefix(&mut contents, Some("> x <- 1\n"), true);
-        let text = match &contents[0] {
-            WorkerContent::ContentText { text, .. } => text.as_str(),
-            _ => "",
-        };
-        assert_eq!(text, "stderr: boom\n");
-    }
-
-    #[test]
-    fn trim_echo_then_append_protocol_warnings_drops_echo_only_multiline_input() {
-        let warning = "late readline result".to_string();
-        let echo = "> x <- 1\n> y <- 2\n";
-        let mut contents = vec![WorkerContent::stdout(echo)];
-
-        trim_echo_then_append_protocol_warnings(
-            &mut contents,
-            Some(echo),
-            false,
-            true,
-            std::slice::from_ref(&warning),
-        );
-
-        assert_eq!(
-            contents,
-            vec![WorkerContent::server_stderr(format!("[repl] {warning}"))]
-        );
-    }
-
-    #[test]
-    fn trim_echo_then_append_protocol_warnings_keeps_output_before_warning() {
-        let warning = "late readline result".to_string();
-        let mut contents = vec![WorkerContent::stdout("> x <- 1\n[1] 1\n")];
-
-        trim_echo_then_append_protocol_warnings(
-            &mut contents,
-            Some("> x <- 1\n"),
-            true,
-            true,
-            std::slice::from_ref(&warning),
-        );
-
-        assert_eq!(
-            contents,
-            vec![
-                WorkerContent::stdout("[1] 1\n"),
-                WorkerContent::server_stderr(format!("[repl] {warning}")),
-            ]
-        );
-    }
-
-    #[test]
-    fn trim_echo_prefix_after_leading_nonstdout_contents_removes_prompt_fallback_echo() {
-        let mut contents = vec![
-            WorkerContent::stderr("stderr: Error: object 'x' not found\n"),
-            WorkerContent::stdout("> x\n"),
-            WorkerContent::stdout("> "),
-        ];
-
-        let trimmed =
-            trim_echo_prefix_after_leading_nonstdout_contents(&mut contents, Some("> x\n"));
-
-        assert!(trimmed, "expected prompt fallback echo to be trimmed");
-        assert_eq!(
-            contents,
-            vec![
-                WorkerContent::stderr("stderr: Error: object 'x' not found\n"),
-                WorkerContent::stdout("> "),
-            ]
-        );
-    }
-
-    #[test]
-    fn trim_decision_applies_to_any_sideband_echo() {
-        let single = vec![echo_event("> ", "1+1\n")];
-        assert!(should_trim_echo_prefix(&single));
-
-        let continuation = vec![echo_event("> ", "1+\n"), echo_event("+ ", "1\n")];
-        assert!(should_trim_echo_prefix(&continuation));
-
-        let multi = vec![echo_event("> ", "1+1\n"), echo_event("> ", "2+2\n")];
-        assert!(should_trim_echo_prefix(&multi));
-
-        let browser = vec![echo_event("Browse[1]> ", "n\n")];
-        assert!(should_trim_echo_prefix(&browser));
-
-        let readline = vec![echo_event("FIRST> ", "alpha\n")];
-        assert!(should_trim_echo_prefix(&readline));
-    }
-
-    #[test]
-    fn collapse_echo_with_attribution_drops_leading_multi_expression_echo_prefix() {
-        let range = OutputRange {
-            start_offset: 0,
-            end_offset: 27,
-            bytes: b"> x <- 1\n> y <- 2\n[1] 2\n> ".to_vec(),
-            events: Vec::new(),
-            text_spans: vec![OutputTextSpan {
-                start_byte: 0,
-                end_byte: 27,
-                is_stderr: false,
-                origin: ContentOrigin::Worker,
-                source: crate::output_capture::OutputTextSource::Ipc,
-            }],
-        };
-
-        let collapsed = collapse_echo_with_attribution(
-            range,
-            &[echo_event("> ", "x <- 1\n"), echo_event("> ", "y <- 2\n")],
-            0,
-            &["> ".to_string()],
-            EchoCollapseMode::CollapseForFinalReply,
-        );
-
-        assert_eq!(
-            String::from_utf8(collapsed.bytes).expect("utf8"),
-            "[1] 2\n> "
-        );
-        assert!(
-            collapsed.events.is_empty(),
-            "did not expect sideband events"
-        );
-        assert_eq!(
-            collapsed.text_spans.len(),
-            1,
-            "expected collapsed output to stay in one stdout span"
-        );
-        assert_eq!(collapsed.text_spans[0].start_byte, 0);
-        assert_eq!(collapsed.text_spans[0].end_byte, 8);
-        assert!(!collapsed.text_spans[0].is_stderr);
-    }
-
-    #[test]
-    fn collapse_echo_with_attribution_drops_leading_echo_prefix_without_separator_newline() {
-        let range = OutputRange {
-            start_offset: 0,
-            end_offset: 42,
-            bytes: b"> xstderr: Error: object 'x' not found\n> ".to_vec(),
-            events: Vec::new(),
-            text_spans: vec![OutputTextSpan {
-                start_byte: 0,
-                end_byte: 42,
-                is_stderr: false,
-                origin: ContentOrigin::Worker,
-                source: crate::output_capture::OutputTextSource::Ipc,
-            }],
-        };
-
-        let collapsed = collapse_echo_with_attribution(
-            range,
-            &[echo_event("> ", "x\n")],
-            0,
-            &["> ".to_string()],
-            EchoCollapseMode::CollapseForFinalReply,
-        );
-
-        assert_eq!(
-            String::from_utf8(collapsed.bytes).expect("utf8"),
-            "stderr: Error: object 'x' not found\n> "
-        );
-        assert!(
-            collapsed.events.is_empty(),
-            "did not expect sideband events"
-        );
-        assert_eq!(
-            collapsed.text_spans.len(),
-            1,
-            "expected collapsed output to stay in one stdout span"
-        );
-        assert_eq!(collapsed.text_spans[0].start_byte, 0);
-        assert_eq!(collapsed.text_spans[0].end_byte, 38);
-        assert!(!collapsed.text_spans[0].is_stderr);
-    }
-
-    #[test]
-    fn trim_matching_echo_event_suffix_from_contents_trims_late_top_level_echo() {
-        let mut contents = vec![WorkerContent::worker_stdout(
-            "> cat(\"TAIL_ONLY\\n\")\nTAIL_ONLY\n> ",
-        )];
-
-        let trimmed = trim_matching_echo_event_suffix_from_contents(
-            &mut contents,
-            &[
-                echo_event("> ", "cat(\"HEAD_ONLY\\n\")\n"),
-                echo_event("> ", "flush.console()\n"),
-                echo_event("> ", "cat(\"TAIL_ONLY\\n\")\n"),
-            ],
-        );
-
-        assert!(trimmed, "expected late top-level echo to be trimmed");
-        assert_eq!(contents_text(&contents), "TAIL_ONLY\n> ");
-    }
-
-    #[test]
-    fn trim_matching_echo_event_suffix_from_contents_keeps_unmatched_prompt_tail() {
-        let mut contents = vec![WorkerContent::worker_stdout("FIRST> alpha\nSECOND> ")];
-
-        let trimmed = trim_matching_echo_event_suffix_from_contents(
-            &mut contents,
-            &[
-                echo_event("FIRST> ", "alpha\n"),
-                echo_event("SECOND> ", "beta\n"),
-            ],
-        );
-
-        assert!(
-            !trimmed,
-            "did not expect partial prompt transcript to be trimmed without an exact match"
-        );
-        assert_eq!(contents_text(&contents), "FIRST> alpha\nSECOND> ");
-    }
-
-    #[test]
-    fn control_prefix_accepts_immediate_tail_without_newline() {
-        let (action, remaining) =
-            split_write_stdin_control_prefix("\u{3}1+1").expect("expected control prefix");
-        assert!(matches!(action, WriteStdinControlAction::Interrupt));
-        assert_eq!(remaining, "1+1");
-    }
-
-    #[test]
-    fn control_prefix_preserves_immediate_newline_tail() {
-        let (action, remaining) =
-            split_write_stdin_control_prefix("\u{4}\nprint(1)").expect("expected control prefix");
-        assert!(matches!(action, WriteStdinControlAction::Restart));
-        assert_eq!(remaining, "\nprint(1)");
-    }
 
     #[test]
     fn send_worker_request_error_preserves_detached_prefix_count() {
@@ -809,1252 +367,6 @@ mod tests {
             "detached-prefix metadata must survive reset until server-side finalization"
         );
         let WorkerReply::Output { .. } = reply;
-    }
-
-    #[test]
-    fn busy_follow_up_reply_sets_busy_error_code_when_missing() {
-        let mut reply = WorkerReply::Output {
-            contents: vec![WorkerContent::worker_stdout("tail\n")],
-            is_error: false,
-            error_code: None,
-            prompt: None,
-            prompt_variants: None,
-        };
-
-        mark_busy_follow_up_reply(&mut reply);
-
-        let WorkerReply::Output {
-            contents,
-            is_error,
-            error_code,
-            ..
-        } = reply;
-        let text = contents
-            .into_iter()
-            .filter_map(|content| match content {
-                WorkerContent::ContentText { text, .. } => Some(text),
-                WorkerContent::ContentImage { .. } => None,
-            })
-            .collect::<String>();
-
-        assert!(
-            is_error,
-            "expected busy follow-up replies to be marked as errors"
-        );
-        assert_eq!(error_code, Some(WorkerErrorCode::Busy));
-        assert!(
-            text.contains("[repl] input discarded while worker busy"),
-            "expected busy follow-up marker, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn busy_follow_up_reply_preserves_timeout_error_code() {
-        let mut reply = WorkerReply::Output {
-            contents: vec![WorkerContent::server_stdout("<<repl status: busy>>\n")],
-            is_error: false,
-            error_code: Some(WorkerErrorCode::Timeout),
-            prompt: None,
-            prompt_variants: None,
-        };
-
-        mark_busy_follow_up_reply(&mut reply);
-
-        let WorkerReply::Output {
-            contents,
-            is_error,
-            error_code,
-            ..
-        } = reply;
-        let text = contents
-            .into_iter()
-            .filter_map(|content| match content {
-                WorkerContent::ContentText { text, .. } => Some(text),
-                WorkerContent::ContentImage { .. } => None,
-            })
-            .collect::<String>();
-
-        assert!(
-            is_error,
-            "expected timed-out busy follow-up replies to be marked as errors"
-        );
-        assert_eq!(
-            error_code,
-            Some(WorkerErrorCode::Timeout),
-            "expected timed-out busy follow-up replies to preserve Timeout"
-        );
-        assert!(
-            text.contains("[repl] input discarded while worker busy"),
-            "expected busy follow-up marker, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn session_end_reset_preserves_detached_prefix_count() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.last_detached_prefix_item_count = 2;
-        manager.session_end_seen = true;
-
-        manager.maybe_reset_after_session_end();
-
-        if let Some(process) = manager.process.take() {
-            let _ = process.kill();
-        }
-
-        assert_eq!(
-            manager.detached_prefix_item_count(),
-            2,
-            "session-end cleanup must preserve detached-prefix metadata until server finalization"
-        );
-    }
-
-    #[cfg(target_family = "unix")]
-    #[test]
-    fn finish_exited_does_not_signal_reaped_root_pid() {
-        let _guard = env_test_mutex().lock().expect("env mutex");
-        let child = successful_test_child();
-        let (result, kills) =
-            capture_recorded_unix_kills(|| test_worker_process(child).finish_exited());
-
-        assert!(
-            result.is_ok(),
-            "expected finish_exited to succeed: {result:?}"
-        );
-        assert!(
-            kills.is_empty(),
-            "did not expect finish_exited to signal an already reaped root pid, got: {kills:?}"
-        );
-    }
-
-    #[cfg(target_family = "unix")]
-    #[test]
-    fn failing_session_end_notice_flushes_partial_stdout_in_files_mode() {
-        let _guard = env_test_mutex().lock().expect("env mutex");
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.pending_output_tape.append_stdout_bytes(&[0xC3]);
-
-        let mut process = test_worker_process(sleeping_test_child());
-        process.set_exit_status_for_test(failing_test_status());
-        manager.process = Some(process);
-
-        manager.note_session_end(true);
-        let formatted = manager.drain_final_formatted_output();
-        let text = contents_text(&formatted.contents);
-
-        if let Some(process) = manager.process.take() {
-            let _ = process.kill();
-        }
-
-        assert!(
-            text.contains("\\xC3"),
-            "expected the partial stdout tail to survive the exit-status notice, got: {text:?}"
-        );
-        assert!(
-            text.contains("worker exited with status 7"),
-            "expected the exit-status notice to stay visible, got: {text:?}"
-        );
-    }
-
-    #[cfg(target_family = "unix")]
-    #[test]
-    fn timed_out_prompt_completion_with_exited_worker_reports_session_end_immediately() {
-        let _guard = env_test_mutex().lock().expect("env mutex");
-        let (server, worker) = crate::ipc::test_connection_pair().expect("ipc pair");
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        let mut process = test_worker_process(successful_test_child());
-        let status = process.wait_child_for_test().expect("wait test child");
-        process.set_exit_status_for_test(status);
-        process.set_ipc_for_test(server);
-        manager.process = Some(process);
-        manager.pending_request = true;
-        manager.pending_request_started_at = Some(std::time::Instant::now());
-        manager.pending_request_input = Some("quit()\n".to_string());
-
-        let prompt = ">>> ".to_string();
-        let _ = worker.send(WorkerToServerIpcMessage::ReadlineStart {
-            prompt: prompt.clone(),
-        });
-        let _ = worker.send(WorkerToServerIpcMessage::ReadlineResult {
-            prompt,
-            line: "quit()\n".to_string(),
-        });
-        let _ = worker.send(WorkerToServerIpcMessage::ReadlineStart {
-            prompt: ">>> ".to_string(),
-        });
-        drop(worker);
-        manager.resolve_timeout_marker_with_wait(Duration::from_millis(200));
-        let formatted = manager.drain_final_formatted_output();
-        let text = contents_text(&formatted.contents);
-
-        assert!(
-            manager.session_end_seen,
-            "expected timed-out completion resolution to notice the exited session"
-        );
-        assert!(
-            manager
-                .settled_pending_completion
-                .as_ref()
-                .is_some_and(|completion| completion.session_end_seen),
-            "expected queued completion metadata to be marked as session-ended"
-        );
-        assert!(
-            text.contains("[repl] session ended"),
-            "expected timed-out completion resolution to record the session-end notice, got: {text:?}"
-        );
-        assert!(
-            !text.contains(">>> "),
-            "did not expect the exited session to keep advertising its prompt, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn files_prepare_input_context_trims_echo_from_prompt_fallback_when_echo_events_missing() {
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(b">>> import time; time.sleep(0.2)\nDETACHED_OK\n");
-        manager.pending_request_input = Some("import time; time.sleep(0.2)\n".to_string());
-        manager.settled_pending_completion = Some(CompletionInfo {
-            prompt: Some(">>> ".to_string()),
-            stdin_wait_prompt: None,
-            prompt_variants: Some(vec![">>> ".to_string()]),
-            echo_events: Vec::new(),
-            protocol_warnings: Vec::new(),
-            session_end_seen: false,
-        });
-
-        let context = manager.prepare_input_context_files();
-        let text = contents_text(&context.detached_prefix_contents);
-
-        assert!(
-            text.contains("DETACHED_OK\n"),
-            "expected the settled files-mode output to survive trimming, got: {text:?}"
-        );
-        assert!(
-            !text.contains("import time; time.sleep(0.2)"),
-            "did not expect the Python prompt echo to leak into the next files-mode reply, got: {text:?}"
-        );
-        assert!(
-            manager.settled_pending_completion.is_none(),
-            "expected settled completion metadata to be consumed with the detached prefix"
-        );
-    }
-
-    #[test]
-    fn interrupt_files_drains_settled_completion_without_leaking_echo() {
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(b">>> import time; time.sleep(0.07)\nDETACHED_OK\n");
-        manager.pending_request_input = Some("import time; time.sleep(0.07)\n".to_string());
-        manager.settled_pending_completion = Some(CompletionInfo {
-            prompt: Some(">>> ".to_string()),
-            stdin_wait_prompt: None,
-            prompt_variants: Some(vec![">>> ".to_string()]),
-            echo_events: Vec::new(),
-            protocol_warnings: Vec::new(),
-            session_end_seen: false,
-        });
-
-        let WorkerReply::Output { contents, .. } = manager
-            .interrupt(Duration::from_millis(10), None, false)
-            .expect("interrupt reply");
-        let text = contents_text(&contents);
-
-        assert!(
-            text.contains("DETACHED_OK\n"),
-            "expected the settled completion output to be preserved, got: {text:?}"
-        );
-        assert!(
-            !text.contains("import time; time.sleep(0.07)"),
-            "did not expect the settled completion echo to leak through interrupt handling, got: {text:?}"
-        );
-        assert!(
-            text.contains(">>> "),
-            "expected the settled completion to keep the prompt on the interrupt reply, got: {text:?}"
-        );
-        assert!(
-            manager.settled_pending_completion.is_none(),
-            "expected the settled completion to be consumed by the interrupt follow-up"
-        );
-    }
-
-    #[test]
-    fn files_empty_poll_waits_for_late_stdout_after_settled_completion() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.process = Some(test_worker_process(sleeping_test_child()));
-        manager.settled_pending_completion = Some(CompletionInfo {
-            prompt: Some("> ".to_string()),
-            stdin_wait_prompt: None,
-            prompt_variants: Some(vec!["> ".to_string()]),
-            echo_events: Vec::new(),
-            protocol_warnings: Vec::new(),
-            session_end_seen: false,
-        });
-
-        let tape = manager.pending_output_tape.clone();
-        let writer = std::thread::spawn(move || {
-            std::thread::sleep(Duration::from_millis(50));
-            tape.append_stdout_bytes(b"[1] 2\n");
-        });
-
-        let reply = manager
-            .write_stdin_files(
-                String::new(),
-                Duration::from_millis(500),
-                Duration::from_millis(500),
-                WriteStdinOptions::default(),
-            )
-            .expect("empty poll reply");
-
-        writer.join().expect("late stdout writer");
-        if let Some(process) = manager.process.take() {
-            let _ = process.kill();
-        }
-
-        let WorkerReply::Output { contents, .. } = reply;
-        let text = contents_text(&contents);
-        assert!(
-            text.contains("[1] 2\n"),
-            "expected the empty poll to wait for late settled stdout, got: {text:?}"
-        );
-        assert!(
-            !text.contains("<<repl status: idle>>"),
-            "did not expect an idle marker before late settled stdout, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn files_reset_preserving_detached_output_keeps_pending_request_input_for_trim() {
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(b">>> import time; time.sleep(0.2)\nDETACHED_OK\n");
-        manager.pending_request_input = Some("import time; time.sleep(0.2)\n".to_string());
-        manager.settled_pending_completion = Some(CompletionInfo {
-            prompt: Some(">>> ".to_string()),
-            stdin_wait_prompt: None,
-            prompt_variants: Some(vec![">>> ".to_string()]),
-            echo_events: Vec::new(),
-            protocol_warnings: Vec::new(),
-            session_end_seen: false,
-        });
-
-        manager.reset_output_state_files_preserving_detached_output();
-
-        let context = manager.prepare_input_context_files();
-        let text = contents_text(&context.detached_prefix_contents);
-
-        assert!(
-            text.contains("DETACHED_OK\n"),
-            "expected detached files-mode output to survive the preserved reset, got: {text:?}"
-        );
-        assert!(
-            !text.contains("import time; time.sleep(0.2)"),
-            "did not expect the preserved reset to leak the original Python input echo, got: {text:?}"
-        );
-        assert!(
-            manager.pending_request_input.is_none(),
-            "expected preserved pending input to be consumed once the detached prefix is prepared"
-        );
-    }
-
-    #[test]
-    fn files_respawned_pending_request_trims_echo_without_settled_completion() {
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.pending_request = true;
-        manager.last_prompt = Some(">>> ".to_string());
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(b">>> import time; time.sleep(0.2)\nDETACHED_OK\n");
-        manager.pending_request_input = Some("import time; time.sleep(0.2)\n".to_string());
-
-        manager.reset_output_state_files_preserving_detached_output();
-
-        let context = manager.prepare_input_context_files();
-        let text = contents_text(&context.detached_prefix_contents);
-
-        assert!(
-            text.contains("DETACHED_OK\n"),
-            "expected aborted pending output to survive the respawned reset, got: {text:?}"
-        );
-        assert!(
-            !text.contains("import time; time.sleep(0.2)"),
-            "did not expect the aborted request echo to leak across the respawn boundary, got: {text:?}"
-        );
-        assert!(
-            manager.pending_request_input.is_none(),
-            "expected the aborted request input fallback to be consumed once the detached prefix is prepared"
-        );
-    }
-
-    #[test]
-    fn pager_respawned_pending_request_trims_echo_without_echo_events() {
-        let _guard = output_ring_test_guard();
-        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
-        reset_output_ring();
-        reset_last_reply_marker_offset();
-
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Pager,
-        )
-        .expect("worker manager");
-        manager.pending_request = true;
-        manager.last_prompt = Some(">>> ".to_string());
-        manager.pending_request_input = Some("import time; time.sleep(0.2)\n".to_string());
-        manager.output.start_capture();
-        manager.output_timeline.append_ipc_text_with_continuation(
-            b">>> import time; time.sleep(0.2)\nDETACHED_OK\n",
-            false,
-            ContentOrigin::Worker,
-            false,
-        );
-
-        manager.reset_output_state_pager_preserving_detached_output(false);
-
-        let context = manager.prepare_input_context_pager("1+1", false);
-        let text = contents_text(&context.detached_prefix_contents);
-
-        assert!(
-            text.contains("DETACHED_OK\n"),
-            "expected aborted pager output to survive the respawned reset, got: {text:?}"
-        );
-        assert!(
-            !text.contains("import time; time.sleep(0.2)"),
-            "did not expect the aborted pager echo to leak across the respawn boundary, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn files_prepare_input_context_seals_split_utf8_at_request_boundary() {
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.pending_output_tape.append_stdout_bytes(&[0xC3]);
-
-        let first = manager.prepare_input_context_files();
-        assert_eq!(
-            contents_text(&first.detached_prefix_contents),
-            "\\xC3",
-            "expected an accepted request to seal the detached utf-8 lead byte into the prefix"
-        );
-
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(&[0xA9, b'\n']);
-        let second = manager.prepare_input_context_files();
-
-        assert_eq!(
-            contents_text(&second.detached_prefix_contents),
-            "\\xA9\n",
-            "expected the next request output to stay split after the detached prefix was sealed"
-        );
-    }
-
-    #[test]
-    fn files_nonfinal_drain_preserves_echo_only_input() {
-        let manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-
-        manager
-            .pending_output_tape
-            .append_stdout_ipc_bytes(b"> Sys.sleep(5)\n");
-        manager
-            .pending_output_tape
-            .append_sideband(PendingSidebandKind::ReadlineResult {
-                prompt: "> ".to_string(),
-                line: "Sys.sleep(5)\n".to_string(),
-                echo_source: PendingTextSource::Ipc,
-            });
-
-        let formatted = manager.drain_formatted_output();
-
-        assert_eq!(
-            formatted.contents,
-            vec![WorkerContent::stdout("> Sys.sleep(5)\n")],
-            "expected an in-flight files-mode drain to keep the echoed command visible"
-        );
-    }
-
-    #[test]
-    fn files_nonfinal_drain_drops_leading_repl_echo_after_worker_output() {
-        let manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-
-        manager
-            .pending_output_tape
-            .append_stdout_ipc_bytes(b"> Sys.sleep(5)\n");
-        manager
-            .pending_output_tape
-            .append_sideband(PendingSidebandKind::ReadlineResult {
-                prompt: "> ".to_string(),
-                line: "Sys.sleep(5)\n".to_string(),
-                echo_source: PendingTextSource::Ipc,
-            });
-        manager.pending_output_tape.append_stdout_bytes(b"start\n");
-
-        let formatted = manager.drain_formatted_output();
-
-        assert_eq!(
-            formatted.contents,
-            vec![WorkerContent::stdout("start\n")],
-            "expected worker output to hide the leading timed-out REPL echo again"
-        );
-    }
-
-    #[test]
-    fn files_prepare_input_context_preserves_unsettled_echo_prefix() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-
-        manager
-            .pending_output_tape
-            .append_stdout_ipc_bytes(b"> Sys.sleep(5)\n");
-        manager
-            .pending_output_tape
-            .append_sideband(PendingSidebandKind::ReadlineResult {
-                prompt: "> ".to_string(),
-                line: "Sys.sleep(5)\n".to_string(),
-                echo_source: PendingTextSource::Ipc,
-            });
-
-        let context = manager.prepare_input_context_files();
-
-        assert_eq!(
-            context.detached_prefix_contents,
-            vec![WorkerContent::stdout("> Sys.sleep(5)\n")],
-            "expected a sealed files-mode prefix without settled completion metadata to keep echoed input"
-        );
-    }
-
-    #[test]
-    fn files_preserved_detached_prefix_stays_separate_from_new_session_startup_output() {
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(b"OLD_TAIL\n");
-
-        manager.reset_output_state_files_preserving_detached_output();
-        manager.next_live_prefix_belongs_to_reply = true;
-        manager
-            .pending_output_tape
-            .append_stdout_bytes(b"NEW_SESSION_STARTUP\n");
-
-        let context = manager.prepare_input_context_files();
-
-        assert_eq!(
-            contents_text(&context.detached_prefix_contents),
-            "OLD_TAIL\n",
-            "expected preserved detached output to stay isolated from the replacement session"
-        );
-        assert_eq!(
-            contents_text(&context.reply_prefix_contents),
-            "NEW_SESSION_STARTUP\n",
-            "expected fresh-session startup output to stay with the new reply prefix"
-        );
-    }
-
-    #[test]
-    fn busy_guardrail_event_survives_sandbox_restart_notice() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.exe_path = PathBuf::from("definitely-missing-worker-exe");
-        manager.stage_sandbox_change_restart_notice(true);
-        manager.guardrail.busy.store(true, Ordering::Relaxed);
-        {
-            let mut slot = manager
-                .guardrail
-                .event
-                .lock()
-                .expect("guardrail event mutex poisoned");
-            *slot = Some(GuardrailEvent {
-                message: "[repl] worker killed by memory guardrail\n".to_string(),
-                was_busy: true,
-                is_error: true,
-            });
-        }
-
-        let reply = manager
-            .write_stdin_files(
-                "1+1".to_string(),
-                Duration::from_millis(10),
-                Duration::from_millis(10),
-                WriteStdinOptions::default(),
-            )
-            .expect("guardrail reply");
-        let WorkerReply::Output { contents, .. } = reply;
-        let text = contents_text(&contents);
-
-        assert!(
-            text.contains("sandbox policy changed; new session started"),
-            "expected the queued restart notice to stay visible, got: {text:?}"
-        );
-        assert!(
-            text.contains("worker error: [repl] worker killed by memory guardrail"),
-            "expected the busy guardrail error to remain authoritative, got: {text:?}"
-        );
-        assert!(
-            !manager.guardrail_busy_event_pending(),
-            "expected the busy guardrail slot to be consumed by the local retry reply"
-        );
-        assert!(
-            manager.pending_server_notice.is_none(),
-            "expected the restart notice to be emitted instead of lingering"
-        );
-        if let Some(process) = manager.process.take() {
-            let _ = process.kill();
-        }
-    }
-
-    #[test]
-    fn bare_restart_flushes_queued_sandbox_change_notice() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.stage_sandbox_change_restart_notice(true);
-
-        let reply = manager
-            .write_stdin_files(
-                "\u{4}".to_string(),
-                Duration::from_millis(10),
-                Duration::from_millis(10),
-                WriteStdinOptions::default(),
-            )
-            .expect("restart reply");
-        let WorkerReply::Output { contents, .. } = reply;
-        let text = contents_text(&contents);
-
-        assert!(
-            text.contains("sandbox policy changed; new session started"),
-            "expected bare restart to flush the queued sandbox notice, got: {text:?}"
-        );
-        assert!(
-            text.contains("[repl] new session started"),
-            "expected the explicit restart notice to remain visible, got: {text:?}"
-        );
-        assert!(
-            manager.pending_server_notice.is_none(),
-            "expected the queued sandbox notice to be consumed by the restart reply"
-        );
-    }
-
-    #[test]
-    fn pager_collapsed_settled_completion_trims_echo_and_keeps_output() {
-        let range = OutputRange {
-            start_offset: 0,
-            end_offset: 25,
-            bytes: b"> Sys.sleep(0.2); 1+1\n[1] 2\n".to_vec(),
-            events: Vec::new(),
-            text_spans: vec![OutputTextSpan {
-                start_byte: 0,
-                end_byte: 25,
-                is_stderr: false,
-                origin: ContentOrigin::Worker,
-                source: crate::output_capture::OutputTextSource::Ipc,
-            }],
-        };
-
-        let collapsed = collapse_echo_with_attribution(
-            range,
-            &[echo_event("> ", "Sys.sleep(0.2); 1+1\n")],
-            0,
-            &["> ".to_string()],
-            EchoCollapseMode::CollapseForFinalReply,
-        );
-        let contents = pager::contents_from_collapsed_output(
-            collapsed.bytes,
-            collapsed.events,
-            collapsed.text_spans,
-            25,
-        );
-        let text = contents_text(&contents);
-
-        assert!(
-            text.contains("[1] 2\n"),
-            "expected settled pager output to be preserved, got: {text:?}"
-        );
-        assert!(
-            !text.contains("Sys.sleep(0.2); 1+1"),
-            "did not expect settled pager echo to leak into the next input context, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn pager_empty_input_polls_pending_output_before_pager_commands() {
-        let _guard = output_ring_test_guard();
-        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
-        reset_output_ring();
-        reset_last_reply_marker_offset();
-
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Pager,
-        )
-        .expect("worker manager");
-        manager.process = Some(test_worker_process(sleeping_test_child()));
-
-        manager.pager.activate(
-            pager_buffer_from_worker_text("line0001\nline0002\nline0003\nline0004\n"),
-            false,
-        );
-
-        manager.output.start_capture();
-        if let Some(end_offset) = manager.output.end_offset() {
-            manager.output.advance_offset_to(end_offset);
-        }
-        manager
-            .output_timeline
-            .append_text(b"detached\n", false, ContentOrigin::Worker);
-
-        let reply = manager
-            .write_stdin_pager(
-                String::new(),
-                Duration::from_millis(0),
-                Duration::from_millis(0),
-                WriteStdinOptions {
-                    page_bytes_override: Some(16),
-                    echo_input: true,
-                    ..WriteStdinOptions::default()
-                },
-            )
-            .expect("empty poll reply");
-
-        let WorkerReply::Output { contents, .. } = reply;
-        let text = contents_text(&contents);
-        assert!(
-            text.contains("detached\n"),
-            "expected empty input to poll newly appended output before pager navigation, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn pager_empty_input_advances_page_after_worker_exit() {
-        let _guard = output_ring_test_guard();
-        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
-        reset_output_ring();
-        reset_last_reply_marker_offset();
-
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Pager,
-        )
-        .expect("worker manager");
-        let mut process = test_worker_process(successful_test_child());
-        let status = process.wait_child_for_test().expect("wait test child");
-        process.set_exit_status_for_test(status);
-        manager.process = Some(process);
-        manager.exe_path = PathBuf::from("definitely-missing-worker-exe");
-
-        let output = (1..=24).map(|n| format!("L{n:04}\n")).collect::<String>();
-        manager
-            .pager
-            .activate(static_pager_buffer_from_worker_text(&output), false);
-        manager.output.start_capture();
-        if let Some(end_offset) = manager.output.end_offset() {
-            manager.output.advance_offset_to(end_offset);
-        }
-        let reply = manager
-            .write_stdin_pager(
-                String::new(),
-                Duration::from_millis(0),
-                Duration::from_millis(0),
-                WriteStdinOptions {
-                    page_bytes_override: Some(16),
-                    echo_input: true,
-                    ..WriteStdinOptions::default()
-                },
-            )
-            .expect("empty pager reply");
-        let WorkerReply::Output { contents, .. } = reply;
-        let text = contents_text(&contents);
-
-        if let Some(process) = manager.process.take() {
-            let _ = process.finish_exited();
-        }
-
-        assert!(
-            text.contains("L0002")
-                || text.contains("L0003")
-                || text.contains("L0010")
-                || text.contains("L0014"),
-            "expected blank pager input to advance to the next page after worker exit, got: {text:?}"
-        );
-        assert!(
-            !text.contains("worker io error:"),
-            "expected pager navigation instead of a respawn error after worker exit, got: {text:?}"
-        );
-    }
-
-    #[test]
-    fn bare_restart_clears_preserved_detached_prefixes() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.preserved_detached_prefix = PrefixCapture {
-            contents: vec![WorkerContent::worker_stdout("OLD_DETACHED\n")],
-            is_error: false,
-            bytes: "OLD_DETACHED\n".len() as u64,
-        };
-        manager.reply_owned_prefix = PrefixCapture {
-            contents: vec![WorkerContent::worker_stdout("OLD_REPLY\n")],
-            is_error: false,
-            bytes: "OLD_REPLY\n".len() as u64,
-        };
-        manager.next_live_prefix_belongs_to_reply = true;
-
-        let reply = manager
-            .write_stdin_files(
-                "\u{4}".to_string(),
-                Duration::from_millis(10),
-                Duration::from_millis(10),
-                WriteStdinOptions::default(),
-            )
-            .expect("restart reply");
-        let WorkerReply::Output { contents, .. } = reply;
-        let reply_text = contents_text(&contents);
-        assert!(
-            !reply_text.contains("OLD_DETACHED") && !reply_text.contains("OLD_REPLY"),
-            "did not expect preserved detached prefixes in restart reply, got: {reply_text:?}"
-        );
-
-        let context = manager.prepare_input_context_files();
-        assert!(
-            context.detached_prefix_contents.is_empty() && context.reply_prefix_contents.is_empty(),
-            "did not expect explicit restart to leak old prefixes into the next input"
-        );
-    }
-
-    #[test]
-    fn pager_empty_input_preserves_idle_guardrail_notice() {
-        let _guard = output_ring_test_guard();
-        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
-
-        let mut last_text = String::new();
-        for _ in 0..16 {
-            reset_output_ring();
-            reset_last_reply_marker_offset();
-
-            let mut manager = WorkerManager::new(
-                Backend::R,
-                SandboxCliPlan::default(),
-                crate::oversized_output::OversizedOutputMode::Pager,
-            )
-            .expect("worker manager");
-            manager.process = Some(test_worker_process(sleeping_test_child()));
-            {
-                let mut slot = manager
-                    .guardrail
-                    .event
-                    .lock()
-                    .expect("guardrail event mutex poisoned");
-                *slot = Some(GuardrailEvent {
-                    message: "[repl] worker was idle; new session started\n".to_string(),
-                    was_busy: false,
-                    is_error: false,
-                });
-            }
-
-            let reply = manager
-                .write_stdin_pager(
-                    String::new(),
-                    Duration::from_millis(0),
-                    Duration::from_millis(0),
-                    WriteStdinOptions {
-                        page_bytes_override: Some(OUTPUT_RING_CAPACITY_BYTES as u64),
-                        echo_input: true,
-                        ..WriteStdinOptions::default()
-                    },
-                )
-                .expect("empty poll reply");
-            let WorkerReply::Output { contents, .. } = reply;
-            last_text = contents_text(&contents);
-
-            if let Some(process) = manager.process.take() {
-                let _ = process.kill();
-            }
-
-            if last_text.contains("[repl] worker was idle; new session started") {
-                return;
-            }
-
-            thread::sleep(Duration::from_millis(5));
-        }
-
-        assert!(
-            last_text.contains("[repl] worker was idle; new session started"),
-            "expected empty pager polls to preserve idle guardrail restart notices, got: {last_text:?}"
-        );
-    }
-
-    #[test]
-    fn pager_output_capture_skips_pending_output_tape() {
-        let output_ring = Arc::new(OutputRing::with_capacity(OUTPUT_RING_CAPACITY_BYTES));
-        let tape = PendingOutputTape::new();
-        let capture = LiveOutputCapture::new(
-            OversizedOutputMode::Pager,
-            tape.clone(),
-            OutputTimeline::new(output_ring.clone()),
-        );
-        capture.append_output_text(b"pager output\n", TextStream::Stdout, false);
-        capture.append_image(IpcPlotImage {
-            id: "img-1".to_string(),
-            data: "AA==".to_string(),
-            mime_type: "image/png".to_string(),
-            is_new: true,
-            updates_previous_image: false,
-            readline_results_seen: 0,
-        });
-        capture.append_sideband(PendingSidebandKind::RequestBoundary);
-
-        assert!(
-            tape.drain_final_snapshot().events.is_empty(),
-            "pager mode should not mirror text, images, or sideband events into the pending tape"
-        );
-        let output_end = output_ring.end_offset();
-        let output_range = output_ring.read_range(0, output_end);
-        assert!(
-            output_end > 0,
-            "pager mode should still append text to the output timeline"
-        );
-        assert_eq!(
-            output_range.bytes, b"pager output\n",
-            "pager mode should keep stdout text in the output timeline"
-        );
-        assert!(
-            output_range.events.iter().any(|event| {
-                matches!(
-                    &event.kind,
-                    OutputEventKind::Image { id, mime_type, .. }
-                    if id == "img-1" && mime_type == "image/png"
-                )
-            }),
-            "pager mode should keep image events in the output timeline"
-        );
-    }
-
-    #[test]
-    fn files_output_capture_anchors_update_notice_before_late_echo() {
-        let output_ring = Arc::new(OutputRing::with_capacity(OUTPUT_RING_CAPACITY_BYTES));
-        let tape = PendingOutputTape::new();
-        let capture = LiveOutputCapture::new(
-            OversizedOutputMode::Files,
-            tape.clone(),
-            OutputTimeline::new(output_ring),
-        );
-
-        capture.append_sideband(PendingSidebandKind::ReadlineResult {
-            prompt: "> ".to_string(),
-            line: "lines(4:8, 4:8)\n".to_string(),
-            echo_source: PendingTextSource::Ipc,
-        });
-        capture.append_image(IpcPlotImage {
-            id: "img-1".to_string(),
-            data: "AA==".to_string(),
-            mime_type: "image/png".to_string(),
-            is_new: true,
-            updates_previous_image: true,
-            readline_results_seen: 1,
-        });
-        capture.append_output_text(b"> lines(4:8, 4:8)\n", TextStream::Stdout, false);
-
-        let contents = tape
-            .drain_final_snapshot()
-            .format_contents_for_reply()
-            .contents;
-
-        assert_eq!(
-            contents,
-            vec![
-                WorkerContent::server_stdout(PREVIOUS_IMAGE_UPDATE_NOTICE),
-                WorkerContent::ContentImage {
-                    data: "AA==".to_string(),
-                    mime_type: "image/png".to_string(),
-                    id: "img-1".to_string(),
-                    is_new: true,
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn files_ipc_output_text_appends_to_tape_and_timeline_in_ipc_order() {
-        let output_ring = Arc::new(OutputRing::with_capacity(OUTPUT_RING_CAPACITY_BYTES));
-        let tape = PendingOutputTape::new();
-        let capture = LiveOutputCapture::new(
-            OversizedOutputMode::Files,
-            tape.clone(),
-            OutputTimeline::new(output_ring.clone()),
-        );
-        let (done_tx, done_rx) = mpsc::channel();
-
-        let output_capture = capture.clone();
-        let start_capture = capture.clone();
-        let result_capture = capture.clone();
-        let image_capture = capture.clone();
-        let session_capture = capture.clone();
-        let (_server, worker) = crate::ipc::test_connection_pair_with_handlers(IpcHandlers {
-            on_output_text: Some(Arc::new(move |text| {
-                output_capture.append_output_text(&text.bytes, text.stream, text.is_continuation);
-            })),
-            on_readline_start: Some(Arc::new(move |prompt| {
-                start_capture.append_sideband(PendingSidebandKind::ReadlineStart { prompt });
-            })),
-            on_readline_result: Some(Arc::new(move |event| {
-                result_capture.append_sideband(PendingSidebandKind::ReadlineResult {
-                    prompt: event.prompt,
-                    line: event.line,
-                    echo_source: PendingTextSource::Ipc,
-                });
-            })),
-            on_plot_image: Some(Arc::new(move |image| {
-                image_capture.append_image(image);
-            })),
-            on_session_end: Some(Arc::new(move || {
-                session_capture.append_sideband(PendingSidebandKind::SessionEnd);
-                done_tx.send(()).expect("send session end marker");
-            })),
-            ..IpcHandlers::default()
-        })
-        .expect("ipc pair");
-
-        worker
-            .send(WorkerToServerIpcMessage::ReadlineStart {
-                prompt: "> ".to_string(),
-            })
-            .expect("send readline_start");
-        worker
-            .send(WorkerToServerIpcMessage::OutputText {
-                stream: TextStream::Stdout,
-                data_b64: base64::engine::general_purpose::STANDARD.encode(b"before\n"),
-                is_continuation: false,
-            })
-            .expect("send stdout output_text");
-        worker
-            .send(WorkerToServerIpcMessage::ReadlineResult {
-                prompt: "> ".to_string(),
-                line: "plot(1)\n".to_string(),
-            })
-            .expect("send readline_result");
-        worker
-            .send(WorkerToServerIpcMessage::PlotImage {
-                mime_type: "image/png".to_string(),
-                data: "AA==".to_string(),
-                is_update: false,
-                source: None,
-            })
-            .expect("send plot_image");
-        worker
-            .send(WorkerToServerIpcMessage::OutputText {
-                stream: TextStream::Stderr,
-                data_b64: base64::engine::general_purpose::STANDARD.encode(b"err\n"),
-                is_continuation: false,
-            })
-            .expect("send stderr output_text");
-        worker
-            .send(WorkerToServerIpcMessage::SessionEnd {
-                reason: None,
-                message_b64: None,
-                turn_id: None,
-            })
-            .expect("send session_end");
-
-        done_rx
-            .recv_timeout(Duration::from_millis(200))
-            .expect("server IPC consumed session_end");
-
-        let snapshot = tape.drain_final_snapshot();
-        assert_eq!(snapshot.events.len(), 6);
-        assert!(matches!(
-            &snapshot.events[0],
-            PendingOutputEvent::Sideband {
-                kind: PendingSidebandKind::ReadlineStart { prompt },
-                ..
-            } if prompt == "> "
-        ));
-        assert!(matches!(
-            &snapshot.events[1],
-            PendingOutputEvent::TextFragment {
-                stream: TextStream::Stdout,
-                origin: ContentOrigin::Worker,
-                bytes,
-                ..
-            } if bytes == b"before\n"
-        ));
-        assert!(matches!(
-            &snapshot.events[2],
-            PendingOutputEvent::Sideband {
-                kind: PendingSidebandKind::ReadlineResult { prompt, line, .. },
-                ..
-            } if prompt == "> " && line == "plot(1)\n"
-        ));
-        assert!(matches!(
-            &snapshot.events[3],
-            PendingOutputEvent::Image {
-                id,
-                mime_type,
-                readline_results_seen: 1,
-                ..
-            } if id.starts_with("image-") && mime_type == "image/png"
-        ));
-        assert!(matches!(
-            &snapshot.events[4],
-            PendingOutputEvent::TextFragment {
-                stream: TextStream::Stderr,
-                origin: ContentOrigin::Worker,
-                bytes,
-                ..
-            } if bytes == b"err\n"
-        ));
-        assert!(matches!(
-            &snapshot.events[5],
-            PendingOutputEvent::Sideband {
-                kind: PendingSidebandKind::SessionEnd,
-                ..
-            }
-        ));
-
-        let end = output_ring.end_offset();
-        let range = output_ring.read_range(0, end);
-        assert_eq!(range.bytes, b"before\n\nstderr: err\n");
-        let image_event = range
-            .events
-            .iter()
-            .find_map(|event| match &event.kind {
-                OutputEventKind::Image {
-                    id,
-                    mime_type,
-                    readline_results_seen,
-                    ..
-                } => Some((event.offset, id, mime_type, readline_results_seen)),
-                _ => None,
-            })
-            .expect("timeline image event");
-        assert_eq!(image_event.0, b"before\n".len() as u64);
-        assert!(image_event.1.starts_with("image-"));
-        assert_eq!(image_event.2, "image/png");
-        assert_eq!(*image_event.3, 1);
-    }
-
-    #[test]
-    fn pager_output_capture_anchors_update_notice_before_late_echo() {
-        let output_ring = Arc::new(OutputRing::with_capacity(OUTPUT_RING_CAPACITY_BYTES));
-        let capture = LiveOutputCapture::new(
-            OversizedOutputMode::Pager,
-            PendingOutputTape::new(),
-            OutputTimeline::new(output_ring.clone()),
-        );
-
-        capture.append_image(IpcPlotImage {
-            id: "img-1".to_string(),
-            data: "AA==".to_string(),
-            mime_type: "image/png".to_string(),
-            is_new: true,
-            updates_previous_image: true,
-            readline_results_seen: 1,
-        });
-        capture.append_output_text(b"> lines(4:8, 4:8)\n", TextStream::Stdout, false);
-
-        let end = output_ring.end_offset();
-        let collapsed = collapse_echo_with_attribution(
-            output_ring.read_range(0, end),
-            &[echo_event("> ", "lines(4:8, 4:8)\n")],
-            0,
-            &["> ".to_string()],
-            EchoCollapseMode::CollapseForFinalReply,
-        );
-        let contents = pager::contents_from_collapsed_output(
-            collapsed.bytes,
-            collapsed.events,
-            collapsed.text_spans,
-            end,
-        );
-
-        assert_eq!(
-            contents,
-            vec![
-                WorkerContent::server_stdout(PREVIOUS_IMAGE_UPDATE_NOTICE),
-                WorkerContent::ContentImage {
-                    data: "AA==".to_string(),
-                    mime_type: "image/png".to_string(),
-                    id: "img-1".to_string(),
-                    is_new: true,
-                },
-            ]
-        );
     }
 
     #[cfg(target_family = "unix")]
@@ -2089,1029 +401,5 @@ mod tests {
         }
 
         assert!(result.is_ok(), "WorkerManager::new should not panic");
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_bwrap_startup_retry_matches_worker_ready_failures() {
-        for message in [
-            "ipc disconnected while waiting for worker_ready",
-            "worker session ended before worker_ready",
-        ] {
-            assert!(
-                linux_sandbox_startup_retryable(&WorkerError::Protocol(message.to_string())),
-                "expected worker_ready startup failure to be retryable: {message}"
-            );
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_bwrap_startup_retry_disables_bwrap_and_announces_fallback() {
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager.sandbox_state.use_linux_sandbox_bwrap = true;
-        manager.sandbox_defaults.use_linux_sandbox_bwrap = true;
-        manager.inherited_sandbox_state = Some(SandboxState {
-            use_linux_sandbox_bwrap: true,
-            ..manager.sandbox_state.clone()
-        });
-
-        let retry = manager.maybe_retry_spawn_without_linux_bwrap(
-            &WorkerError::Protocol("ipc disconnected while waiting for backend info".to_string()),
-            false,
-        );
-
-        assert!(
-            retry,
-            "expected backend-info disconnect to trigger bwrap fallback"
-        );
-        assert!(
-            !manager.sandbox_state.use_linux_sandbox_bwrap,
-            "expected effective sandbox state to disable bwrap after fallback"
-        );
-        assert!(
-            !manager.sandbox_defaults.use_linux_sandbox_bwrap,
-            "expected sandbox defaults to disable bwrap after fallback"
-        );
-        assert!(
-            manager
-                .inherited_sandbox_state
-                .as_ref()
-                .is_some_and(|state| !state.use_linux_sandbox_bwrap),
-            "expected inherited sandbox state to disable bwrap after fallback"
-        );
-
-        let snapshot = manager.pending_output_tape.drain_final_snapshot();
-        let text = contents_text(&snapshot.format_contents().contents);
-        assert!(
-            text.contains("continuing without bwrap"),
-            "expected fallback notice in visible output, got: {text:?}"
-        );
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_bwrap_startup_retry_stays_disabled_after_followup_codex_meta_update() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        let mut inherited_state = manager.sandbox_defaults.clone();
-        inherited_state.apply_update(SandboxStateUpdate {
-            sandbox_policy: SandboxPolicy::WorkspaceWrite {
-                writable_roots: Vec::new(),
-                network_access: false,
-                exclude_tmpdir_env_var: false,
-                exclude_slash_tmp: false,
-            },
-            sandbox_cwd: Some(std::env::temp_dir()),
-            use_linux_sandbox_bwrap: Some(true),
-            use_legacy_landlock: None,
-        });
-        manager.inherited_sandbox_state = Some(inherited_state.clone());
-        manager.sandbox_state = resolve_effective_sandbox_state_with_defaults(
-            &manager.sandbox_plan,
-            Some(&inherited_state),
-            &manager.sandbox_defaults,
-        )
-        .expect("resolved initial sandbox state");
-        assert!(
-            manager.sandbox_state.use_linux_sandbox_bwrap,
-            "test setup should start with bwrap enabled"
-        );
-
-        let retry = manager.maybe_retry_spawn_without_linux_bwrap(
-            &WorkerError::Protocol("ipc disconnected while waiting for backend info".to_string()),
-            false,
-        );
-        assert!(retry, "expected startup failure to disable bwrap");
-
-        let update = sandbox_state_update_from_codex_meta(&json!({
-            "sandboxPolicy": {
-                "type": "workspace-write",
-                "writable_roots": [],
-                "network_access": false,
-                "exclude_tmpdir_env_var": false,
-                "exclude_slash_tmp": false
-            },
-            "sandboxCwd": std::env::temp_dir(),
-            "useLegacyLandlock": false,
-            "codexLinuxSandboxExe": "/tmp/codex-linux-sandbox"
-        }))
-        .expect("Codex sandbox metadata");
-        manager
-            .update_sandbox_state(update, Duration::from_millis(1))
-            .expect("follow-up sandbox state");
-
-        assert!(
-            !manager.sandbox_state.use_linux_sandbox_bwrap,
-            "follow-up Codex metadata should preserve the local no-bwrap fallback"
-        );
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn linux_bwrap_startup_retry_stays_disabled_after_followup_plan_bwrap_override() {
-        let plan = SandboxCliPlan {
-            operations: vec![
-                crate::sandbox_cli::SandboxCliOperation::SetMode(
-                    crate::sandbox_cli::SandboxModeArg::Inherit,
-                ),
-                crate::sandbox_cli::SandboxCliOperation::Config(
-                    crate::sandbox_cli::SandboxConfigOperation::SetUseLinuxSandboxBwrap(true),
-                ),
-            ],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        let mut inherited_state = manager.sandbox_defaults.clone();
-        inherited_state.apply_update(SandboxStateUpdate {
-            sandbox_policy: SandboxPolicy::WorkspaceWrite {
-                writable_roots: Vec::new(),
-                network_access: false,
-                exclude_tmpdir_env_var: false,
-                exclude_slash_tmp: false,
-            },
-            sandbox_cwd: Some(std::env::temp_dir()),
-            use_linux_sandbox_bwrap: None,
-            use_legacy_landlock: None,
-        });
-        manager.inherited_sandbox_state = Some(inherited_state.clone());
-        manager.sandbox_state = resolve_effective_sandbox_state_with_defaults(
-            &manager.sandbox_plan,
-            Some(&inherited_state),
-            &manager.sandbox_defaults,
-        )
-        .expect("resolved initial sandbox state");
-        assert!(
-            manager.sandbox_state.use_linux_sandbox_bwrap,
-            "test setup should start with the plan-level bwrap override enabled"
-        );
-
-        let retry = manager.maybe_retry_spawn_without_linux_bwrap(
-            &WorkerError::Protocol("ipc disconnected while waiting for backend info".to_string()),
-            false,
-        );
-        assert!(retry, "expected startup failure to disable bwrap");
-
-        let update = sandbox_state_update_from_codex_meta(&json!({
-            "sandboxPolicy": {
-                "type": "workspace-write",
-                "writable_roots": [],
-                "network_access": false,
-                "exclude_tmpdir_env_var": false,
-                "exclude_slash_tmp": false
-            },
-            "sandboxCwd": std::env::temp_dir(),
-            "useLegacyLandlock": false,
-            "codexLinuxSandboxExe": "/tmp/codex-linux-sandbox"
-        }))
-        .expect("Codex sandbox metadata");
-        manager
-            .update_sandbox_state(update, Duration::from_millis(1))
-            .expect("follow-up sandbox state");
-
-        assert!(
-            !manager.sandbox_state.use_linux_sandbox_bwrap,
-            "plan-level bwrap overrides should not re-enable bwrap after the local fallback"
-        );
-    }
-
-    #[test]
-    fn inherit_ending_invalid_plan_fails_during_startup_validation() {
-        let plan = SandboxCliPlan {
-            operations: vec![
-                crate::sandbox_cli::SandboxCliOperation::SetMode(
-                    crate::sandbox_cli::SandboxModeArg::ReadOnly,
-                ),
-                crate::sandbox_cli::SandboxCliOperation::AddWritableRoot(std::env::temp_dir()),
-                crate::sandbox_cli::SandboxCliOperation::SetMode(
-                    crate::sandbox_cli::SandboxModeArg::Inherit,
-                ),
-            ],
-        };
-
-        let err = match WorkerManager::new(
-            Backend::Python,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        ) {
-            Ok(_) => panic!("invalid inherit-ending plan should fail during startup"),
-            Err(err) => err,
-        };
-
-        assert!(
-            matches!(err, WorkerError::Sandbox(ref message) if message.contains("--add-writable-root can only be used while sandbox mode is workspace-write")),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn inherit_workspace_write_refinements_wait_for_client_state() {
-        let writable_root = std::env::temp_dir();
-        let plan = SandboxCliPlan {
-            operations: vec![
-                crate::sandbox_cli::SandboxCliOperation::SetMode(
-                    crate::sandbox_cli::SandboxModeArg::Inherit,
-                ),
-                crate::sandbox_cli::SandboxCliOperation::AddWritableRoot(writable_root.clone()),
-                crate::sandbox_cli::SandboxCliOperation::Config(
-                    crate::sandbox_cli::SandboxConfigOperation::SetWorkspaceNetworkAccess(true),
-                ),
-            ],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-
-        manager
-            .stage_sandbox_state_update(SandboxStateUpdate {
-                sandbox_policy: SandboxPolicy::WorkspaceWrite {
-                    writable_roots: Vec::new(),
-                    network_access: false,
-                    exclude_tmpdir_env_var: false,
-                    exclude_slash_tmp: false,
-                },
-                sandbox_cwd: Some(writable_root.clone()),
-                use_linux_sandbox_bwrap: None,
-                use_legacy_landlock: None,
-            })
-            .expect("workspace-write Codex metadata should satisfy deferred refinements");
-
-        let SandboxPolicy::WorkspaceWrite {
-            writable_roots,
-            network_access,
-            ..
-        } = &manager.sandbox_state.sandbox_policy
-        else {
-            panic!(
-                "expected staged inherit refinements to resolve to workspace-write, got {:?}",
-                manager.sandbox_state.sandbox_policy
-            );
-        };
-        assert!(
-            *network_access,
-            "expected deferred workspace network setting to apply after client metadata"
-        );
-        assert!(
-            writable_roots.iter().any(|path| path == &writable_root),
-            "expected deferred writable root to apply after client metadata"
-        );
-    }
-
-    #[test]
-    fn failed_sandbox_update_does_not_commit_inherited_state() {
-        let _guard = cwd_test_mutex().lock().expect("cwd mutex");
-        let plan = crate::sandbox_cli::SandboxCliPlan {
-            operations: vec![
-                crate::sandbox_cli::SandboxCliOperation::SetMode(
-                    crate::sandbox_cli::SandboxModeArg::Inherit,
-                ),
-                crate::sandbox_cli::SandboxCliOperation::Config(
-                    crate::sandbox_cli::SandboxConfigOperation::SetWorkspaceNetworkAccess(true),
-                ),
-            ],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        let mut inherited_before = manager.sandbox_defaults.clone();
-        inherited_before.apply_update(SandboxStateUpdate {
-            sandbox_policy: SandboxPolicy::WorkspaceWrite {
-                writable_roots: Vec::new(),
-                network_access: false,
-                exclude_tmpdir_env_var: false,
-                exclude_slash_tmp: false,
-            },
-            sandbox_cwd: None,
-            use_linux_sandbox_bwrap: None,
-            use_legacy_landlock: None,
-        });
-        manager.inherited_sandbox_state = Some(inherited_before.clone());
-        manager.sandbox_state = resolve_effective_sandbox_state_with_defaults(
-            &manager.sandbox_plan,
-            Some(&inherited_before),
-            &manager.sandbox_defaults,
-        )
-        .expect("resolved initial inherited sandbox state");
-
-        let err = manager
-            .update_sandbox_state(
-                SandboxStateUpdate {
-                    sandbox_policy: SandboxPolicy::DangerFullAccess,
-                    sandbox_cwd: None,
-                    use_linux_sandbox_bwrap: None,
-                    use_legacy_landlock: None,
-                },
-                Duration::from_millis(1),
-            )
-            .expect_err("danger-full-access should fail workspace-write-only config");
-        assert!(
-            matches!(err, WorkerError::Sandbox(ref msg) if msg.contains("requires workspace-write mode")),
-            "unexpected error: {err}"
-        );
-        assert_eq!(
-            manager.inherited_sandbox_state,
-            Some(inherited_before),
-            "failed updates must not mutate inherited sandbox baseline"
-        );
-    }
-
-    #[test]
-    fn exact_interrupt_remains_local_when_worker_would_respawn() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-
-        assert!(
-            manager
-                .nonexecuting_follow_up_uses_existing_state("\u{3}")
-                .expect("interrupt follow-up classification"),
-            "a bare Ctrl-C should stay a local follow-up even when it would otherwise respawn"
-        );
-    }
-
-    #[test]
-    fn interrupt_pager_tail_requires_current_sandbox_when_worker_would_respawn() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Pager,
-        )
-        .expect("worker manager");
-        let mut process = test_worker_process(successful_test_child());
-        process
-            .wait_child_for_test()
-            .expect("wait for the stub worker process to exit");
-        manager.process = Some(process);
-
-        manager.output.start_capture();
-        manager.output_timeline.append_text(
-            b"line0001\nline0002\nline0003\nline0004\n",
-            false,
-            ContentOrigin::Worker,
-        );
-        let end_offset = manager.output.end_offset().expect("output end offset");
-        let SnapshotWithImages { buffer, .. } =
-            snapshot_page_with_images(&manager.output, end_offset, 16);
-        manager.pager.activate(buffer.expect("pager buffer"), false);
-
-        assert!(
-            !manager
-                .nonexecuting_follow_up_uses_existing_state("\u{3}:q")
-                .expect("interrupt follow-up classification"),
-            "a pager ctrl-c tail should require current per-call sandbox metadata when it would respawn"
-        );
-    }
-
-    #[test]
-    fn empty_input_with_busy_guardrail_uses_existing_state() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        {
-            let mut slot = manager
-                .guardrail
-                .event
-                .lock()
-                .expect("guardrail event mutex poisoned");
-            *slot = Some(GuardrailEvent {
-                message: "[repl] previous request aborted; retry your last input\n".to_string(),
-                was_busy: true,
-                is_error: true,
-            });
-        }
-
-        assert!(
-            !manager
-                .empty_input_requires_spawn()
-                .expect("empty-input classification"),
-            "empty polls should keep pending busy-guardrail recovery local"
-        );
-    }
-
-    #[test]
-    fn nonempty_input_with_busy_guardrail_requires_current_state() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        {
-            let mut slot = manager
-                .guardrail
-                .event
-                .lock()
-                .expect("guardrail event mutex poisoned");
-            *slot = Some(GuardrailEvent {
-                message: "[repl] previous request aborted; retry your last input\n".to_string(),
-                was_busy: true,
-                is_error: true,
-            });
-        }
-
-        assert!(
-            !manager
-                .nonexecuting_follow_up_uses_existing_state("1+1")
-                .expect("follow-up classification"),
-            "busy-guardrail retries should require current per-call sandbox metadata"
-        );
-    }
-
-    #[test]
-    fn empty_input_with_idle_guardrail_requires_spawn() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        {
-            let mut slot = manager
-                .guardrail
-                .event
-                .lock()
-                .expect("guardrail event mutex poisoned");
-            *slot = Some(GuardrailEvent {
-                message: "[repl] worker was idle; new session started\n".to_string(),
-                was_busy: false,
-                is_error: false,
-            });
-        }
-
-        assert!(
-            manager
-                .empty_input_requires_spawn()
-                .expect("empty-input classification"),
-            "idle guardrail notices should still require current per-call sandbox metadata when a poll would respawn"
-        );
-    }
-
-    #[test]
-    fn prechecked_empty_input_requires_current_sandbox_when_worker_exited() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .stage_sandbox_state_update(SandboxStateUpdate {
-                sandbox_policy: SandboxPolicy::ReadOnly,
-                sandbox_cwd: None,
-                use_linux_sandbox_bwrap: None,
-                use_legacy_landlock: None,
-            })
-            .expect("initial inherited state");
-        let mut process = test_worker_process(successful_test_child());
-        process
-            .wait_child_for_test()
-            .expect("wait for the stub worker process to exit");
-        manager.process = Some(process);
-
-        let result = manager.write_stdin(
-            String::new(),
-            Duration::from_secs(1),
-            Duration::from_secs(1),
-            WriteStdinOptions {
-                pending_state_prechecked: true,
-                ..WriteStdinOptions::default()
-            },
-        );
-
-        assert!(
-            matches!(result, Err(ref err) if is_prechecked_follow_up_requires_meta(err)),
-            "expected prechecked empty input to require current sandbox metadata once the worker has exited, got: {result:?}"
-        );
-    }
-
-    #[test]
-    fn prechecked_bare_interrupt_requires_current_sandbox_when_worker_exited() {
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .stage_sandbox_state_update(SandboxStateUpdate {
-                sandbox_policy: SandboxPolicy::ReadOnly,
-                sandbox_cwd: None,
-                use_linux_sandbox_bwrap: None,
-                use_legacy_landlock: None,
-            })
-            .expect("initial inherited state");
-        let mut process = test_worker_process(successful_test_child());
-        process
-            .wait_child_for_test()
-            .expect("wait for the stub worker process to exit");
-        manager.process = Some(process);
-
-        let result = manager.write_stdin(
-            "\u{3}".to_string(),
-            Duration::from_secs(1),
-            Duration::from_secs(1),
-            WriteStdinOptions {
-                pending_state_prechecked: true,
-                ..WriteStdinOptions::default()
-            },
-        );
-
-        assert!(
-            matches!(result, Err(ref err) if is_prechecked_follow_up_requires_meta(err)),
-            "expected prechecked bare ctrl-c to require current sandbox metadata once the worker has exited, got: {result:?}"
-        );
-    }
-
-    #[cfg(target_family = "unix")]
-    #[test]
-    fn interrupt_tail_uses_current_sandbox_for_the_respawn() {
-        let _guard = cwd_test_mutex().lock().expect("cwd mutex");
-        let temp = tempfile::Builder::new()
-            .prefix(".tmp-interrupt-tail-current-sandbox-")
-            .tempdir_in(worker_process_test_temp_parent("worker-process"))
-            .expect("tempdir");
-        let sandbox_cwd = temp.path().to_path_buf();
-        let plan = SandboxCliPlan {
-            operations: vec![crate::sandbox_cli::SandboxCliOperation::SetMode(
-                crate::sandbox_cli::SandboxModeArg::Inherit,
-            )],
-        };
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            plan,
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        manager
-            .stage_sandbox_state_update(SandboxStateUpdate {
-                sandbox_policy: SandboxPolicy::ReadOnly,
-                sandbox_cwd: Some(sandbox_cwd.clone()),
-                use_linux_sandbox_bwrap: None,
-                use_legacy_landlock: None,
-            })
-            .expect("initial inherited read-only state");
-        let mut process = test_worker_process(successful_test_child());
-        process
-            .wait_child_for_test()
-            .expect("wait for the stub worker process to exit");
-        manager.process = Some(process);
-        manager.exe_path = PathBuf::from("definitely-missing-worker-exe");
-
-        let result = manager.write_stdin(
-            "\u{3}1+1".to_string(),
-            Duration::from_secs(10),
-            Duration::from_secs(10),
-            WriteStdinOptions {
-                deferred_sandbox_state_update: Some(SandboxStateUpdate {
-                    sandbox_policy: SandboxPolicy::WorkspaceWrite {
-                        writable_roots: Vec::new(),
-                        network_access: false,
-                        exclude_tmpdir_env_var: false,
-                        exclude_slash_tmp: false,
-                    },
-                    sandbox_cwd: Some(sandbox_cwd.clone()),
-                    use_linux_sandbox_bwrap: None,
-                    use_legacy_landlock: None,
-                }),
-                ..WriteStdinOptions::default()
-            },
-        );
-        match result {
-            Ok(WorkerReply::Output {
-                contents, is_error, ..
-            }) => {
-                let text = contents_text(&contents);
-                assert!(
-                    is_error,
-                    "expected the failed interrupt-tail respawn attempt to surface as an error reply"
-                );
-                assert!(
-                    text.contains("worker error:"),
-                    "expected the failed interrupt-tail respawn attempt to report a worker error, got: {text:?}"
-                );
-            }
-            Err(WorkerError::Protocol(message)) => {
-                assert!(
-                    message.contains("backend info") || message.contains("ipc disconnected"),
-                    "expected the failed interrupt-tail respawn attempt to fail during worker startup, got: {message:?}"
-                );
-            }
-            Err(err) => panic!("unexpected interrupt-tail respawn error: {err}"),
-        }
-        assert!(
-            matches!(
-                manager.sandbox_state.sandbox_policy,
-                SandboxPolicy::WorkspaceWrite { .. }
-            ),
-            "expected deferred metadata to stage before interrupt attempts the respawn"
-        );
-        assert_eq!(
-            manager.sandbox_state.sandbox_cwd, sandbox_cwd,
-            "expected deferred metadata to update the effective sandbox cwd before the respawn path"
-        );
-    }
-
-    #[test]
-    fn apply_debug_startup_env_uses_mcp_repl_vars() {
-        let _guard = env_test_mutex().lock().expect("env mutex");
-        let original = std::env::var_os(crate::debug_logs::DEBUG_SESSION_DIR_ENV);
-        let original_startup_path = std::env::var_os(crate::diagnostics::STARTUP_LOG_PATH_ENV);
-        unsafe {
-            std::env::set_var(
-                crate::debug_logs::DEBUG_SESSION_DIR_ENV,
-                "/tmp/mcp-repl-debug-session",
-            );
-            std::env::remove_var(crate::diagnostics::STARTUP_LOG_PATH_ENV);
-        }
-
-        let mut command = Command::new("env");
-        apply_debug_startup_env(&mut command, None);
-        let envs: std::collections::BTreeMap<_, _> = command
-            .get_envs()
-            .map(|(key, value)| {
-                (
-                    key.to_string_lossy().into_owned(),
-                    value.map(|value| value.to_string_lossy().into_owned()),
-                )
-            })
-            .collect();
-
-        match original {
-            Some(value) => unsafe {
-                std::env::set_var(crate::debug_logs::DEBUG_SESSION_DIR_ENV, value);
-            },
-            None => unsafe {
-                std::env::remove_var(crate::debug_logs::DEBUG_SESSION_DIR_ENV);
-            },
-        }
-        match original_startup_path {
-            Some(value) => unsafe {
-                std::env::set_var(crate::diagnostics::STARTUP_LOG_PATH_ENV, value);
-            },
-            None => unsafe {
-                std::env::remove_var(crate::diagnostics::STARTUP_LOG_PATH_ENV);
-            },
-        }
-
-        assert_eq!(
-            envs.get(crate::debug_logs::DEBUG_SESSION_DIR_ENV),
-            Some(&Some("/tmp/mcp-repl-debug-session".to_string()))
-        );
-        assert_eq!(envs.get(crate::diagnostics::STARTUP_LOG_PATH_ENV), None);
-    }
-
-    #[test]
-    fn apply_debug_startup_env_uses_session_tmpdir_for_worker_log() {
-        let _guard = env_test_mutex().lock().expect("env mutex");
-        let original = std::env::var_os(crate::debug_logs::DEBUG_SESSION_DIR_ENV);
-        let original_startup_path = std::env::var_os(crate::diagnostics::STARTUP_LOG_PATH_ENV);
-        unsafe {
-            std::env::set_var(
-                crate::debug_logs::DEBUG_SESSION_DIR_ENV,
-                "/tmp/mcp-repl-debug-session",
-            );
-            std::env::remove_var(crate::diagnostics::STARTUP_LOG_PATH_ENV);
-        }
-
-        let mut command = Command::new("env");
-        let session_tmpdir = PathBuf::from("/tmp/mcp-repl-session-tmp");
-        apply_debug_startup_env(&mut command, Some(&session_tmpdir));
-        let envs: std::collections::BTreeMap<_, _> = command
-            .get_envs()
-            .map(|(key, value)| {
-                (
-                    key.to_string_lossy().into_owned(),
-                    value.map(|value| value.to_string_lossy().into_owned()),
-                )
-            })
-            .collect();
-
-        match original {
-            Some(value) => unsafe {
-                std::env::set_var(crate::debug_logs::DEBUG_SESSION_DIR_ENV, value);
-            },
-            None => unsafe {
-                std::env::remove_var(crate::debug_logs::DEBUG_SESSION_DIR_ENV);
-            },
-        }
-        match original_startup_path {
-            Some(value) => unsafe {
-                std::env::set_var(crate::diagnostics::STARTUP_LOG_PATH_ENV, value);
-            },
-            None => unsafe {
-                std::env::remove_var(crate::diagnostics::STARTUP_LOG_PATH_ENV);
-            },
-        }
-
-        assert_eq!(
-            envs.get(crate::debug_logs::DEBUG_SESSION_DIR_ENV),
-            Some(&Some("/tmp/mcp-repl-debug-session".to_string()))
-        );
-        assert_eq!(
-            envs.get(crate::diagnostics::STARTUP_LOG_PATH_ENV),
-            Some(&Some(
-                session_tmpdir
-                    .join(crate::diagnostics::WORKER_STARTUP_LOG_FILE_NAME)
-                    .display()
-                    .to_string()
-            ))
-        );
-    }
-
-    #[test]
-    fn persist_worker_startup_log_copies_into_debug_session_dir() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let session_tmpdir = temp.path().join("session-tmp");
-        let debug_session_dir = temp.path().join("debug-session");
-        std::fs::create_dir_all(&session_tmpdir).expect("create session tmpdir");
-        std::fs::create_dir_all(&debug_session_dir).expect("create debug session dir");
-
-        let source = session_tmpdir.join(crate::diagnostics::WORKER_STARTUP_LOG_FILE_NAME);
-        let destination = debug_session_dir.join(crate::diagnostics::WORKER_STARTUP_LOG_FILE_NAME);
-        std::fs::write(&source, "worker startup log\n").expect("write source log");
-
-        persist_worker_startup_log(&session_tmpdir, Some(destination.clone()));
-
-        assert_eq!(
-            std::fs::read_to_string(&destination).expect("read destination log"),
-            "worker startup log\n"
-        );
-    }
-
-    #[test]
-    fn cleanup_worker_session_tmpdir_persists_log_when_keep_tmpdir_is_set() {
-        let _guard = env_test_mutex().lock().expect("env mutex");
-        let temp = tempfile::tempdir().expect("tempdir");
-        let session_tmpdir = temp.path().join("session-tmp");
-        let debug_session_dir = temp.path().join("debug-session");
-        std::fs::create_dir_all(&session_tmpdir).expect("create session tmpdir");
-        std::fs::create_dir_all(&debug_session_dir).expect("create debug session dir");
-
-        let source = session_tmpdir.join(crate::diagnostics::WORKER_STARTUP_LOG_FILE_NAME);
-        let destination = debug_session_dir.join(crate::diagnostics::WORKER_STARTUP_LOG_FILE_NAME);
-        std::fs::write(&source, "worker startup log\n").expect("write source log");
-
-        let original_keep = std::env::var_os("MCP_REPL_KEEP_SESSION_TMPDIR");
-        unsafe {
-            std::env::set_var("MCP_REPL_KEEP_SESSION_TMPDIR", "1");
-        }
-
-        cleanup_worker_session_tmpdir(&session_tmpdir, Some(destination.clone()));
-
-        match original_keep {
-            Some(value) => unsafe {
-                std::env::set_var("MCP_REPL_KEEP_SESSION_TMPDIR", value);
-            },
-            None => unsafe {
-                std::env::remove_var("MCP_REPL_KEEP_SESSION_TMPDIR");
-            },
-        }
-
-        assert!(
-            session_tmpdir.is_dir(),
-            "session tmpdir should be preserved"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&destination).expect("read destination log"),
-            "worker startup log\n"
-        );
-    }
-
-    #[cfg(target_family = "windows")]
-    #[test]
-    fn windows_ipc_connect_error_reaps_wrapper_process() {
-        let mut child = Command::new("powershell.exe")
-            .args(["-NoProfile", "-Command", "Start-Sleep -Seconds 30"])
-            .spawn()
-            .expect("spawn test child process");
-
-        let result = handle_windows_ipc_connect_result(
-            Err(std::io::Error::other("ipc connect failed")),
-            &mut child,
-        );
-        assert!(matches!(result, Err(WorkerError::Io(_))));
-
-        let deadline = std::time::Instant::now() + Duration::from_secs(2);
-        loop {
-            let status = child.try_wait().expect("query child status");
-            if status.is_some() {
-                break;
-            }
-            if std::time::Instant::now() >= deadline {
-                let _ = child.kill();
-                let _ = child.wait();
-                panic!("connect-error handler should reap child wrapper");
-            }
-            std::thread::sleep(Duration::from_millis(20));
-        }
-    }
-
-    #[cfg(target_family = "windows")]
-    #[test]
-    fn windows_soft_termination_does_not_kill_child() {
-        let mut child = Command::new("powershell.exe")
-            .args(["-NoProfile", "-Command", "Start-Sleep -Seconds 30"])
-            .spawn()
-            .expect("spawn test child process");
-
-        request_soft_termination(&mut child).expect("soft terminate call should succeed");
-
-        let status = child.try_wait().expect("query child status");
-        assert!(
-            status.is_none(),
-            "child should still be running after soft termination request"
-        );
-
-        let _ = child.kill();
-        let _ = child.wait();
-    }
-
-    #[cfg(target_family = "windows")]
-    #[test]
-    fn windows_ipc_connect_timeout_is_bounded() {
-        assert!(
-            WINDOWS_IPC_CONNECT_MAX_WAIT <= Duration::from_secs(10),
-            "windows IPC connect max wait should fail fast, got {:?}",
-            WINDOWS_IPC_CONNECT_MAX_WAIT
-        );
-    }
-
-    #[cfg(target_family = "windows")]
-    #[test]
-    fn windows_sandbox_prepare_access_denied_fails_fast() {
-        let _guard = crate::windows_sandbox::prepare_sandbox_launch_test_mutex()
-            .lock()
-            .expect("windows sandbox test mutex");
-        crate::windows_sandbox::set_prepare_sandbox_launch_test_error(Some(
-            "failed to prepare writable ACL on 'C:\\workspace': SetNamedSecurityInfoW failed: 5"
-                .to_string(),
-        ));
-
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        let result = manager.ensure_windows_sandbox_launch();
-
-        crate::windows_sandbox::set_prepare_sandbox_launch_test_error(None);
-
-        assert!(
-            matches!(
-                result,
-                Err(WorkerError::Sandbox(ref message))
-                    if message.contains("SetNamedSecurityInfoW failed: 5")
-            ),
-            "access-denied prepare failures should abort launch preparation, got: {result:?}"
-        );
-        assert!(
-            manager.windows_sandbox_launch.is_none(),
-            "failed launch preparation should not cache a prepared launch"
-        );
-    }
-
-    #[cfg(target_family = "windows")]
-    #[test]
-    fn windows_python_sandbox_prepare_access_denied_fails_fast() {
-        let _guard = crate::windows_sandbox::prepare_sandbox_launch_test_mutex()
-            .lock()
-            .expect("windows sandbox test mutex");
-        crate::windows_sandbox::set_prepare_sandbox_launch_test_error(Some(
-            "failed to prepare writable ACL on 'C:\\workspace': SetNamedSecurityInfoW failed: 5"
-                .to_string(),
-        ));
-
-        let mut manager = WorkerManager::new(
-            Backend::Python,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-        let result = manager.ensure_windows_sandbox_launch();
-
-        crate::windows_sandbox::set_prepare_sandbox_launch_test_error(None);
-
-        assert!(
-            matches!(
-                result,
-                Err(WorkerError::Sandbox(ref message))
-                    if message.contains("SetNamedSecurityInfoW failed: 5")
-            ),
-            "Python access-denied prepare failures should abort launch preparation, got: {result:?}"
-        );
-        assert!(
-            manager.windows_sandbox_launch.is_none(),
-            "failed Python launch preparation should not cache a prepared launch"
-        );
-    }
-
-    #[cfg(target_family = "windows")]
-    #[test]
-    fn windows_sandbox_cache_hit_refreshes_prepared_launch_acl_state_before_reuse() {
-        let _guard = crate::windows_sandbox::prepare_sandbox_launch_test_mutex()
-            .lock()
-            .expect("windows sandbox test mutex");
-
-        let mut manager = WorkerManager::new(
-            Backend::R,
-            SandboxCliPlan::default(),
-            crate::oversized_output::OversizedOutputMode::Files,
-        )
-        .expect("worker manager");
-
-        let first = manager
-            .ensure_windows_sandbox_launch()
-            .expect("initial launch preparation should succeed");
-        assert!(
-            first.is_some(),
-            "initial launch preparation should populate the prepared-launch cache"
-        );
-
-        crate::windows_sandbox::set_apply_prepared_launch_acl_state_test_error(Some(
-            "cache hit should refresh ACL state".to_string(),
-        ));
-
-        let second = manager.ensure_windows_sandbox_launch();
-
-        crate::windows_sandbox::set_apply_prepared_launch_acl_state_test_error(None);
-
-        assert!(
-            matches!(
-                second,
-                Err(WorkerError::Sandbox(ref err))
-                    if err.contains("cache hit should refresh ACL state")
-            ),
-            "cache hits should refresh ACL state before reusing the prepared launch, got: {second:?}"
-        );
-        assert_eq!(
-            manager.windows_sandbox_launch, first,
-            "cache-hit refresh failures should preserve the cached launch for later retries"
-        );
     }
 }
