+++ src/worker_process/write_preflight.rs
@@ -321,3 +321,249 @@ impl WorkerManager {
         Ok(WritePreflightOutcome::Reply(reply))
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::backend::Backend;
+    use crate::output_capture::{
+        OUTPUT_RING_CAPACITY_BYTES, ensure_output_ring, reset_last_reply_marker_offset,
+        reset_output_ring,
+    };
+    use crate::oversized_output::OversizedOutputMode;
+    use crate::sandbox_cli::SandboxCliPlan;
+    use crate::worker_process::test_support::{
+        contents_text, output_ring_test_guard, pager_buffer_from_worker_text, sleeping_test_child,
+        static_pager_buffer_from_worker_text, successful_test_child, test_worker_process,
+    };
+    use crate::worker_protocol::{ContentOrigin, WorkerReply};
+    use crate::worker_supervisor::GuardrailEvent;
+    use std::path::PathBuf;
+    use std::sync::atomic::Ordering;
+
+    #[test]
+    fn busy_guardrail_event_survives_sandbox_restart_notice() {
+        let mut manager = WorkerManager::new(
+            Backend::R,
+            SandboxCliPlan::default(),
+            OversizedOutputMode::Files,
+        )
+        .expect("worker manager");
+        manager.exe_path = PathBuf::from("definitely-missing-worker-exe");
+        manager.stage_sandbox_change_restart_notice(true);
+        manager.guardrail.busy.store(true, Ordering::Relaxed);
+        {
+            let mut slot = manager
+                .guardrail
+                .event
+                .lock()
+                .expect("guardrail event mutex poisoned");
+            *slot = Some(GuardrailEvent {
+                message: "[repl] worker killed by memory guardrail\n".to_string(),
+                was_busy: true,
+                is_error: true,
+            });
+        }
+
+        let reply = manager
+            .write_stdin_files(
+                "1+1".to_string(),
+                Duration::from_millis(10),
+                Duration::from_millis(10),
+                WriteStdinOptions::default(),
+            )
+            .expect("guardrail reply");
+        let WorkerReply::Output { contents, .. } = reply;
+        let text = contents_text(&contents);
+
+        assert!(
+            text.contains("sandbox policy changed; new session started"),
+            "expected the queued restart notice to stay visible, got: {text:?}"
+        );
+        assert!(
+            text.contains("worker error: [repl] worker killed by memory guardrail"),
+            "expected the busy guardrail error to remain authoritative, got: {text:?}"
+        );
+        assert!(
+            !manager.guardrail_busy_event_pending(),
+            "expected the busy guardrail slot to be consumed by the local retry reply"
+        );
+        assert!(
+            manager.pending_server_notice.is_none(),
+            "expected the restart notice to be emitted instead of lingering"
+        );
+        if let Some(process) = manager.process.take() {
+            let _ = process.kill();
+        }
+    }
+
+    #[test]
+    fn pager_empty_input_polls_pending_output_before_pager_commands() {
+        let _guard = output_ring_test_guard();
+        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
+        reset_output_ring();
+        reset_last_reply_marker_offset();
+
+        let mut manager = WorkerManager::new(
+            Backend::R,
+            SandboxCliPlan::default(),
+            OversizedOutputMode::Pager,
+        )
+        .expect("worker manager");
+        manager.process = Some(test_worker_process(sleeping_test_child()));
+
+        manager.pager.activate(
+            pager_buffer_from_worker_text("line0001\nline0002\nline0003\nline0004\n"),
+            false,
+        );
+
+        manager.output.start_capture();
+        if let Some(end_offset) = manager.output.end_offset() {
+            manager.output.advance_offset_to(end_offset);
+        }
+        manager
+            .output_timeline
+            .append_text(b"detached\n", false, ContentOrigin::Worker);
+
+        let reply = manager
+            .write_stdin_pager(
+                String::new(),
+                Duration::from_millis(0),
+                Duration::from_millis(0),
+                WriteStdinOptions {
+                    page_bytes_override: Some(16),
+                    echo_input: true,
+                    ..WriteStdinOptions::default()
+                },
+            )
+            .expect("empty poll reply");
+
+        let WorkerReply::Output { contents, .. } = reply;
+        let text = contents_text(&contents);
+        assert!(
+            text.contains("detached\n"),
+            "expected empty input to poll newly appended output before pager navigation, got: {text:?}"
+        );
+    }
+
+    #[test]
+    fn pager_empty_input_advances_page_after_worker_exit() {
+        let _guard = output_ring_test_guard();
+        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
+        reset_output_ring();
+        reset_last_reply_marker_offset();
+
+        let mut manager = WorkerManager::new(
+            Backend::R,
+            SandboxCliPlan::default(),
+            OversizedOutputMode::Pager,
+        )
+        .expect("worker manager");
+        let mut process = test_worker_process(successful_test_child());
+        let status = process.wait_child_for_test().expect("wait test child");
+        process.set_exit_status_for_test(status);
+        manager.process = Some(process);
+        manager.exe_path = PathBuf::from("definitely-missing-worker-exe");
+
+        let output = (1..=24).map(|n| format!("L{n:04}\n")).collect::<String>();
+        manager
+            .pager
+            .activate(static_pager_buffer_from_worker_text(&output), false);
+        manager.output.start_capture();
+        if let Some(end_offset) = manager.output.end_offset() {
+            manager.output.advance_offset_to(end_offset);
+        }
+        let reply = manager
+            .write_stdin_pager(
+                String::new(),
+                Duration::from_millis(0),
+                Duration::from_millis(0),
+                WriteStdinOptions {
+                    page_bytes_override: Some(16),
+                    echo_input: true,
+                    ..WriteStdinOptions::default()
+                },
+            )
+            .expect("empty pager reply");
+        let WorkerReply::Output { contents, .. } = reply;
+        let text = contents_text(&contents);
+
+        if let Some(process) = manager.process.take() {
+            let _ = process.finish_exited();
+        }
+
+        assert!(
+            text.contains("L0002")
+                || text.contains("L0003")
+                || text.contains("L0010")
+                || text.contains("L0014"),
+            "expected blank pager input to advance to the next page after worker exit, got: {text:?}"
+        );
+        assert!(
+            !text.contains("worker io error:"),
+            "expected pager navigation instead of a respawn error after worker exit, got: {text:?}"
+        );
+    }
+
+    #[test]
+    fn pager_empty_input_preserves_idle_guardrail_notice() {
+        let _guard = output_ring_test_guard();
+        let _output_ring = ensure_output_ring(OUTPUT_RING_CAPACITY_BYTES);
+
+        let mut last_text = String::new();
+        for _ in 0..16 {
+            reset_output_ring();
+            reset_last_reply_marker_offset();
+
+            let mut manager = WorkerManager::new(
+                Backend::R,
+                SandboxCliPlan::default(),
+                OversizedOutputMode::Pager,
+            )
+            .expect("worker manager");
+            manager.process = Some(test_worker_process(sleeping_test_child()));
+            {
+                let mut slot = manager
+                    .guardrail
+                    .event
+                    .lock()
+                    .expect("guardrail event mutex poisoned");
+                *slot = Some(GuardrailEvent {
+                    message: "[repl] worker was idle; new session started\n".to_string(),
+                    was_busy: false,
+                    is_error: false,
+                });
+            }
+
+            let reply = manager
+                .write_stdin_pager(
+                    String::new(),
+                    Duration::from_millis(0),
+                    Duration::from_millis(0),
+                    WriteStdinOptions {
+                        page_bytes_override: Some(OUTPUT_RING_CAPACITY_BYTES as u64),
+                        echo_input: true,
+                        ..WriteStdinOptions::default()
+                    },
+                )
+                .expect("empty poll reply");
+            let WorkerReply::Output { contents, .. } = reply;
+            last_text = contents_text(&contents);
+
+            if let Some(process) = manager.process.take() {
+                let _ = process.kill();
+            }
+
+            if last_text.contains("[repl] worker was idle; new session started") {
+                return;
+            }
+
+            std::thread::sleep(Duration::from_millis(5));
+        }
+
+        assert!(
+            last_text.contains("[repl] worker was idle; new session started"),
+            "expected empty pager polls to preserve idle guardrail restart notices, got: {last_text:?}"
+        );
+    }
+}
